rgw: move RGWPostRESTResourceCR to rgw_cr_rest.h
for use by RGWRealmPusher
